mandatory well tested and managed memory leaks

+++ ft_printf/ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: oessaoud <oessaoud@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/19 23:23:45 by oessaoud          #+#    #+#             */
-/*   Updated: 2024/11/21 22:06:08 by oessaoud         ###   ########.fr       */
+/*   Updated: 2024/12/02 16:58:12 by oessaoud         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,9 +39,9 @@ int	ft_witchformat(char current, va_list lst)
 	else if (current == 'p')
 		count += ft_putaddress_printf(va_arg(lst, void *));
 	else if (current == 'x')
-		count += ft_puthexa_printf(va_arg(lst, unsigned int), 0);
+		count += ft_puthexa_printf(va_arg(lst, unsigned long), 0);
 	else if (current == 'X')
-		count += ft_puthexa_printf(va_arg(lst, unsigned int), 1);
+		count += ft_puthexa_printf(va_arg(lst, unsigned long), 1);
 	else if (current == '%')
 		count += ft_putpourcent_printf();
 	if (count < 0)
@@ -76,17 +76,3 @@ int	ft_printf(const char *str, ...)
 	va_end(args);
 	return (count);
 }
-
-// int main()
-// {
-//  char  *a = "oussa";
-//  //unsigned long adr = (unsigned long)a;
-// 	// int i =ft_printf("helloworold! \n%p ma", a);
-// 	// printf("\n%d\n",i);
-// 	// printf("\n/////////////////\n");
-// 	// int i2 =printf("helloworold! \n%p ma", a);
-// 	int i = ft_printf("\n%u\n",-200000);
-// 	printf("\n%d\n",i);
-// 	i = printf("\n%u\n",-200000);
-// 	printf("\n%d\n",i);
-// }
